updates event component functionality and initiates build
